scripts/requirements.txt
requests==2.31.0
pandas==2.2.0
pyyaml==6.0.1
streamlit>=1.32.0
numpy>=1.26.0
